# specs/004-institution-iam-auth-integration/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Clinica Digital IAM API
  version: 0.3.0
  description: Contratos de IAM nativo in-app, registro institucional, sessao opaca, contexto do Shell e validacao RNDS para Organization, Location, Practitioner e PractitionerRole.
servers:
  - url: http://localhost:8080
paths:
...
  /api/admin/sessions:
    get:
      summary: Lista sessoes administrativas do tenant
      operationId: listAdminSessions
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Lista de sessoes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminSessionSummary'
  /api/admin/sessions/{sessionId}/revoke:
    post:
      summary: Revoga sessao administrativa com motivo auditavel
      operationId: revokeAdminSession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeAdminSessionRequest'
      responses:
        '204':
          description: Sessao revogada
components:
  schemas:
...
    AdminSessionSummary:
      type: object
      required: [id, tenantId, userId, active]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
          nullable: true
        issuedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
        traceId:
          type: string
          nullable: true
        clientIp:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        revocationReason:
          type: string
          nullable: true
        active:
          type: boolean
    RevokeAdminSessionRequest:
      type: object
      properties:
        revocationReason:
          type: string
          maxLength: 64
...